created own file with test cases for quick-fixes #1309
